LPS-55928 rename and move to highlight package
