+++ modules/apps/search/search-web/src/META-INF/resources/init.jsp
@@ -52,7 +52,7 @@ page import="com.liferay.portal.kernel.search.facet.ScopeFacet" %><%@
 page import="com.liferay.portal.kernel.search.facet.collector.FacetCollector" %><%@
 page import="com.liferay.portal.kernel.search.facet.collector.TermCollector" %><%@
 page import="com.liferay.portal.kernel.search.facet.config.FacetConfiguration" %><%@
-page import="com.liferay.portal.kernel.search.util.SearchUtil" %><%@
+page import="com.liferay.portal.kernel.search.highlight.SearchUtil" %><%@
 page import="com.liferay.portal.kernel.util.ArrayUtil" %><%@
 page import="com.liferay.portal.kernel.util.CalendarFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
