+++ portal-service/src/com/liferay/portal/kernel/search/highlight/packageinfo
